Fix units of memory limits in LSFLogParser.
It was using the wrong units for the memory limits.
It was using KB instead of MB.

Also added the ability to specify the unit for the memory limits.

Usage:
./LSFLogParser.py --logfile-dir /path/to/logfiles --output-csv /path/to/output.csv --unit-for-limits MB --default-max-mem-gb 10.0

The default is MB, as specified by LSF, but it can be specified based on how you have it configured in lsf.conf.